Fix unable duplicate html and deprecate the html delete icon button (#910)
* deprecated delete icon button above the html block, because we already have front menu

* remove some unused codes

* fix: unable copy html block
