+++ src/muya/lib/utils/exportMarkdown.js
@@ -249,7 +249,7 @@ class ExportMarkdown {
 
   normalizeHTML (block, indent) { // figure
     const result = []
-    const codeLines = block.children[1].children[0].children[0].children
+    const codeLines = block.children[0].children[0].children[0].children
     for (const line of codeLines) {
       result.push(`${indent}${line.text}\n`)
     }
